Add tests for {HDF5,SimTel}EventSouce filling Provenance info
