+++ src/ctapipe/io/tests/test_hdf5eventsource.py
@@ -270,3 +270,18 @@ def test_simulated_events_distribution(dl1_file):
         dist = source.simulated_shower_distributions[1]
         assert dist["n_entries"] == 1000
         assert dist["histogram"].sum() == 1000.0
+
+
+def test_provenance(dl1_file, provenance):
+    """Make sure that HDF5EventSource reads reference metadata and adds to provenance"""
+    from ctapipe.io.metadata import _read_reference_metadata_hdf5
+
+    provenance.start_activity("test_hdf5eventsource")
+    with HDF5EventSource(input_url=dl1_file):
+        pass
+
+    inputs = provenance.current_activity.input
+    assert len(inputs) == 1
+    assert inputs[0]["url"] == str(dl1_file)
+    meta = _read_reference_metadata_hdf5(dl1_file)
+    assert inputs[0]["reference_meta"].product.id_ == meta.product.id_
